rename test package to ppdd
