+++ src/main/java/ppdd/ClassIShouldFindOne.java
@@ -1,4 +1,4 @@
-package test;
+package ppdd;
 
 public class ClassIShouldFindOne {
 	public static void foo() {
