+++ src/main/java/ppdd/ClassIShouldFindTwo.java
@@ -1,4 +1,4 @@
-package test;
+package ppdd;
 
 public class ClassIShouldFindTwo {
 	public static void foo() {
